created RegistrationController
