+++ src/main/webapp/WEB-INF/view/home.jsp
@@ -1,6 +1,8 @@
 <%@ taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
 <%@ taglib uri="http://www.springframework.org/security/tags"
 	prefix="security"%>
+<link rel="stylesheet"
+	href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
 
 <html>
 
@@ -21,11 +23,9 @@
 	<p>
 		Access for:
 		<security:authentication property="principal.username" />
-		<br> <br> 
-		Your pass:
-		<security:authentication property="principal.password" />		
-		<br> <br> 
-		Your role(s):
+		<br> <br> Your pass:
+		<security:authentication property="principal.password" />
+		<br> <br> Your role(s):
 		<security:authentication property="principal.authorities" />
 
 	</p>
@@ -35,39 +35,49 @@
 	<!--  Add a link to a point to /leaders... this id for the managers -->
 
 	<security:authorize access="hasRole('PATIENT')">
-	<p>
-		<a href="${pageContext.request.contextPath}/patient">
-		Go to Patient Zone
-		</a>
-			(read medical examinations)
-	</p>
+		<p>
+			<a href="${pageContext.request.contextPath}/patient"
+				class="btn btn-primary" role="button" aria-pressed="true"> Go to
+				Patient Zone </a> (read medical examinations)
+		</p>
 
-	<hr>
+		<hr>
 	</security:authorize>
-	
+
 	<!--  Add a link to a point to /leaders... this id for the managers -->
 
 	<security:authorize access="hasRole('DOCTOR')">
-	<p>
-		<a href="${pageContext.request.contextPath}/doctor/">
-		Go to Doctor Zone
-		</a>
-			(create medical examinations)
-	</p>
+		<p>
+			<a href="${pageContext.request.contextPath}/doctor/"
+				class="btn btn-primary" role="button" aria-pressed="true"> Go to
+				Doctor Zone </a> (create medical examinations)
+		</p>
 
-	<hr>
+		<hr>
 	</security:authorize>
-	
+
 
 	<security:authorize access="hasAnyRole('ADMIN', 'MODERATOR')">
 
 		<p>
-			<a href="${pageContext.request.contextPath}/admin/">
-			Go to Admin/Mod Zone
-			</a> (delete medical examination) - Only for the Boss
+			<a href="${pageContext.request.contextPath}/admin/"
+				class="btn btn-primary" role="button" aria-pressed="true"> Go to
+				Admin/Mod Zone </a> (delete medical examination) - Only for the Boss
 		</p>
 
-		
+
+	</security:authorize>
+
+	<hr> 
+	<security:authorize access="hasAnyRole('DOCTOR', 'ADMIN', 'MODERATOR')">
+
+		<p>
+			<a href="${pageContext.request.contextPath}/registration/"
+				class="btn btn-primary" role="button" aria-pressed="true">
+				Register new user </a>
+		</p>
+
+
 	</security:authorize>
 	<!-- Add a logout button -->
 	<form:form action="${pageContext.request.contextPath}/logout"
